LayoutTests/accessibility/children-changed-sends-notification-expected.txt

This test ensures that a notification is being emitted when children are added or removed for an accessibility object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Plain text paragraph

PARAGRAPH notification: AXChildrenAdded
GLOBAL notification: AXChildrenAdded on element with role AXRole: AXParagraph
PARAGRAPH notification: AXChildrenRemoved
GLOBAL notification: AXChildrenRemoved on element with role AXRole: AXParagraph
PASS paragraphNotificationCount is globalNotificationCount
PASS globalNotificationCount is 2
PASS successfullyParsed is true

TEST COMPLETE
